Standardized usage of test key
